training on military aircraft dataset

+++ src/dataset/ingestion/image_ingestion/jpeg_ingester.py
@@ -1,6 +1,7 @@
 from torchvision.io import read_image
+import torch
 
 class JpegIngester:
     @classmethod
     def read(cls, image_path):
-        return read_image(image_path)
+        return read_image(image_path).to(dtype=torch.float32)
