+++ tests/dataset/ingestion/image_ingestion/test_jpeg_ingester.py
@@ -5,7 +5,7 @@ def test_read():
     image_path = "tests/dataset/ingestion/image_ingestion/jpg_fixture.jpg"
     t = JpegIngester.read(image_path)
     assert t.shape == (3, 831, 859)
-    assert t.dtype == torch.uint8
+    assert t.dtype == torch.float32
     
 if __name__ == "__main__":
     test_read()
